Added three screens: Splash Screen, Login/Signup Screen and Profile Setup Screen.
The Splash Screen needs 1 or 2 changes in the UI, more specifially in the UI button.
The background of the Splash Screen is static, as I haven't recieved the GIF yet.
The Login/Signup UI is done. Bloc is yet to be implemented.
The Profile Setup UI is done. Bloc is yet to be implemented.