Added a shimmer loading screen for the profile screen. (#5)
* Added three screens: Splash Screen, Login/Signup Screen and Profile Setup Screen.

The Splash Screen needs 1 or 2 changes in the UI, more specifially in the UI button.
The background of the Splash Screen is static, as I haven't recieved the GIF yet.
The Login/Signup UI is done. Bloc is yet to be implemented.
The Profile Setup UI is done. Bloc is yet to be implemented.

* Added signup/login with email functionality

* Added Google signin functionality

* Added the rules screen

* All the screens are done.
All the routes are done

* All routes and screens are implemented.
Error handling, animations and caching is still left to be done.

* Added .DS_Store

* Added a shimmer animation for the profile screen

* Added the bottom status bar on the questions screen.
Fixed a navigation bug which let users navigate back to the login screen without logging out.

* Added a fading animation while authenticating with Google

* Added a fading placeholder animation for the leaderboards screen.
FIXED BUG: JSON response didnt contain "usedHint" for new users causing the app to break. Fixed it by checking for key first.